Finally a commit, Sex
195 changed files sounds impressive... but it's not.
i also strated ifdefing literally every change made here compared to vanilla so that we can easily find that shit later(some general fixes and things werent documented and will probably be forgotten)

+++ mp/src/game/client/c_baseplayer.h
@@ -270,8 +270,8 @@ public:
 #endif
 	#endif //SecobMod__USE_PLAYERCLASSES
 #ifdef MFS
-	float	m_BlueTime;
-	float	m_RedTime;
+	int	m_BlueTime;
+	int	m_RedTime;
 	int GetBlueTime() { return m_BlueTime; }
 	int GetRedTime() { return m_RedTime; }
 #endif
